backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
celery==5.4.0
redis==5.2.0
google-generativeai==0.8.3
gTTS==2.5.4
mutagen==1.47.0
requests==2.32.3
python-multipart==0.0.20
aiofiles==24.1.0
python-dotenv==1.0.1
faster-whisper==1.1.1

